Updated loading weights script
