+++ benchmarks/model/Benchmark with Hugging Face.csv
@@ -1,6 +1,6 @@
 batch_size,Triton,Huggingface
-1.0,69.0,4.5
-8.0,69.3,21.1
-16.0,69.6,39.3
-32.0,98.0,87.3
-64.0,207.3,151.5
+1.0,9.3,4.4
+8.0,19.6,22.4
+16.0,39.4,40.6
+32.0,89.0,78.1
+64.0,198.8,152.5
