+++ vit/utils.py
@@ -3,14 +3,9 @@ import torch
 import functools
 import numpy as np
 from tqdm import tqdm
-
 from typing import List, Tuple
 
 from loguru import logger
-from .load_weights import (
-    map_non_attn_layers,
-    map_attn_layers
-)
 
 logger.remove()
 logger.add(sys.stdout, format="[{time: YYYY-MM-DD HH:mm:ss} {level}] {message}", level="INFO")
@@ -42,76 +37,6 @@ def tensor_info(func_name):
     return decorator
 
 
-def transfer_pretrained_weights(pretrained_model: torch.nn.Module, custom_model: torch.nn.Module) -> torch.nn.Module:
-    """
-    Transfer weights from a pretrained model to the custom model
-    """
-
-    pretrained_state_dict = pretrained_model.state_dict()
-    custom_state_dict = custom_model.state_dict()
-
-    num_layers = 12
-
-    # Mapping dictionary from source model to destination model
-    weight_mapping = {
-        'embeddings.cls_token': 'embeddings.cls_token',
-        'embeddings.position_embeddings': 'embeddings.position_embeddings',
-        'embeddings.patch_embeddings.projection.weight': 'embeddings.projection.weight',
-        'embeddings.patch_embeddings.projection.bias': 'embeddings.projection.bias',
-        'layernorm.weight': 'layernorm.weight',
-        'layernorm.bias': 'layernorm.bias',
-        'pooler.dense.weight': 'pooler.dense.weight',
-        'pooler.dense.bias': 'pooler.dense.bias'
-    }
-   
-    # Adding mappings for each encoder layer's output and intermediate dense layers
-    for i in range(num_layers):
-        weight_mapping.update({
-            f'encoder.layer.{i}.output.dense.weight': f'encoder.layer.{i}.output.weight',
-            f'encoder.layer.{i}.output.dense.bias': f'encoder.layer.{i}.output.bias',
-            f'encoder.layer.{i}.intermediate.dense.weight': f'encoder.layer.{i}.intermediate.weight',
-            f'encoder.layer.{i}.intermediate.dense.bias': f'encoder.layer.{i}.intermediate.bias',
-            f'encoder.layer.{i}.attention.output.dense.weight': f'encoder.layer.{i}.attention.output.weight',
-            f'encoder.layer.{i}.attention.output.dense.bias': f'encoder.layer.{i}.attention.output.bias',
-            f'encoder.layer.{i}.layernorm_before.weight': f'encoder.layer.{i}.layernorm_before.weight',
-            f'encoder.layer.{i}.layernorm_before.bias': f'encoder.layer.{i}.layernorm_before.bias',
-            f'encoder.layer.{i}.layernorm_after.weight': f'encoder.layer.{i}.layernorm_after.weight',
-            f'encoder.layer.{i}.layernorm_after.bias': f'encoder.layer.{i}.layernorm_after.bias'
-        })
-
-    # Transfer the weights of query, key, value layers
-    attention_layers = [
-        f'encoder.layer.{k}.attention.attention' for k in range(num_layers)
-    ]
-
-    for layer_name, weight in pretrained_state_dict.items():
-        for attn_layer in attention_layers:
-            if attn_layer in layer_name:
-                layer_num, proj, type = layer_name.split('.')[2], layer_name.split('.')[-2], layer_name.split('.')[-1]
-                logger.debug(f'Mapping from source\t{layer_name}\tLayer number: {layer_num} \tProjection: {proj} \tType: {type}')
-                custom_state_dict = map_attn_layers(layer_num, proj, type, weight, custom_state_dict)
-
-    # Transfer rest of the layers
-    custom_state_dict = map_non_attn_layers(
-        source_state_dict=pretrained_state_dict,
-        dest_state_dict=custom_state_dict,
-        weight_mapping=weight_mapping
-    )
-
-    custom_model.load_state_dict(custom_state_dict, strict=False)
-
-    custom_state_dict = custom_model.state_dict()
-    uninitialized_layers = []
-    for k, v in custom_state_dict.items():
-        if torch.all(v == 0):
-            uninitialized_layers.append(k)
-
-    if uninitialized_layers:
-        # Pooler bias has all zeros, so that is expected to be here
-        print(f"Some layer are not initialized: {uninitialized_layers}")
-
-    return custom_model
-
 def capture_cuda_graph(model, static_input):
     stream = torch.cuda.Stream()
     stream.wait_stream(torch.cuda.current_stream())
@@ -140,7 +65,7 @@ def benchmark(
         batch_sizes: List[int] = [1, 4, 16, 32, 64, 128, 256],
         warmups: int = 50,
         reps: int = 100
-    ) -> None:
+    ):
     """
     Benchmark two models on different batch sizes
